normalize naming - ledger router, key image store
