+++ fog/ledger/server/src/lib.rs
@@ -5,8 +5,8 @@ mod config;
 mod counters;
 mod db_fetcher;
 mod error;
-mod key_image_router_server;
-mod key_image_router_service;
+mod router_server;
+mod router_service;
 mod key_image_service;
 mod key_image_store_server;
 mod merkle_proof_service;
